inicial firebase code
