+++ requirements.txt
@@ -39,6 +39,7 @@ faiss-cpu==1.9.0.post1
 fastembed==0.5.1
 filelock==3.17.0
 filetype==1.2.0
+firebase-admin==6.6.0
 Flask==3.1.0
 flatbuffers==25.1.21
 fonttools==4.55.3
